Revert "First push of integration tests. WIP - some passing, some not."
This reverts commit dd2b2a57550300c75d1f16691dedc849a18dd850

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/model/Transfer.java
@@ -12,7 +12,7 @@ public class Transfer {
     //To use for mapRowToTransfer in JdbcTransferDao
     public Transfer() {};
 
-    public Transfer(double amount, int accountToId, int accountFromId, int transferTypeId){
+    public Transfer(int amount, int accountToId, int accountFromId, int transferTypeId){
         this.amount = amount;
         this.accountToId = accountToId;
         this.accountFromId = accountFromId;
